Develop

+++ api/schedules/log_rotate.py
@@ -15,6 +15,9 @@ LOG_ROTATE_CONFIGS = [
 ]
 
 def execute():
+    blockchain = os.environ['blockchains'][0]
+    if blockchain == 'mmx':
+        return  # Only Chia+forks for now
     app.logger.info("Executing log rotation...")
     for config in LOG_ROTATE_CONFIGS:
         if os.path.exists(LOG_ROTATE_CONFIG_DIR + config):
@@ -22,9 +25,10 @@ def execute():
             subprocess.call("/usr/sbin/logrotate " + LOG_ROTATE_CONFIG_DIR + config + " >/dev/null 2>&1", shell=True)
 
     # Extra guards for farmr which can eat GBs of log space sometimes
-    for file in os.listdir("/root/.chia/farmr"):
-        if file.startswith("log"):
-            size_mbs = os.path.getsize(os.path.join("/root/.chia/farmr", file)) >> 20
-            if (size_mbs > MAX_LOG_SIZE_MB): 
-                app.logger.info("Deleting large farmr log at {0}".format(os.path.join("/root/.chia/farmr", file)))
-                os.unlink(os.path.join("/root/.chia/farmr", file))
+    if os.path.exists("/root/.chia/farmr"):
+        for file in os.listdir("/root/.chia/farmr"):
+            if file.startswith("log"):
+                size_mbs = os.path.getsize(os.path.join("/root/.chia/farmr", file)) >> 20
+                if (size_mbs > MAX_LOG_SIZE_MB): 
+                    app.logger.info("Deleting large farmr log at {0}".format(os.path.join("/root/.chia/farmr", file)))
+                    os.unlink(os.path.join("/root/.chia/farmr", file))
